add commandline flag
